make seprate file for theme
